+++ api/src/app.ts
@@ -4,11 +4,12 @@ import morgan from "morgan";
 import cors from "cors";
 import mainRouter from "./routes";
 import { orangeText, syncDatabase } from "./db";
-import { authenticateToken } from "./utils/authenticateToken";
 import loginRouter from "./routes/login";
+import { authenticateToken } from "./utils/authenticateToken";
 
 const PORT = process.env.PORT || 3000;
 const app = express();
+const NODE_ENV = process.env.NODE_ENV || "development";
 
 app.use(morgan("dev"));
 app.use(express.json());
@@ -26,7 +27,8 @@ app.use((_req, res, next) => {
 });
 
 app.use("/login", loginRouter);
-app.use("/", authenticateToken, mainRouter);
+if (NODE_ENV === "production") app.use("/", authenticateToken ,mainRouter);
+if (NODE_ENV === "development") app.use("/", mainRouter);
 
 syncDatabase()
   .then(() => {
