+++ api/src/config.ts
@@ -1,0 +1,15 @@
+export const {
+  // Afip
+  RAZON_SOCIAL,
+  CUIT,
+  DOMICILIO_FISCAL,
+  INICIO_ACTIVIDAD,
+  REGIMEN_TRIBUTARIO,
+  IIBB,
+  // Database
+  NODE_ENV = "development",
+  DB_URL,
+  DB_USER,
+  DB_PASSWORD,
+  DB_HOST,
+} = process.env;
