+++ api/src/utils/afipPDF/facturaA.html
@@ -1,0 +1,150 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <title>Ticket</title>
+    <style type="text/css">
+      * {
+        box-sizing: border-box;
+        -webkit-user-select: none;
+        /* Chrome, Opera, Safari */
+        -moz-user-select: none;
+        /* Firefox 2+ */
+        -ms-user-select: none;
+        /* IE 10+ */
+        user-select: none;
+        /* Standard syntax */
+      }
+
+      .bill-container {
+        border-collapse: collapse;
+        max-width: 8cm;
+        position: absolute;
+        left: 0;
+        right: 0;
+        margin: auto;
+        border-collapse: collapse;
+        font-family: monospace;
+        font-size: 12px;
+      }
+
+      .text-lg {
+        font-size: 20px;
+      }
+
+      .text-center {
+        text-align: center;
+      }
+
+      #qrcode {
+        width: 200px;
+      }
+
+      p {
+        margin: 2px 0;
+      }
+
+      table table {
+        width: 100%;
+      }
+
+      table table tr td:last-child {
+        text-align: right;
+      }
+
+      .border-top {
+        border-top: 1px dashed;
+      }
+
+      .padding-b-3 {
+        padding-bottom: 3px;
+      }
+
+      .padding-t-3 {
+        padding-top: 3px;
+      }
+    </style>
+  </head>
+
+  <body>
+    <table class="bill-container">
+      <tr>
+        <td class="padding-b-3">
+          <p>Razón social: {{razonSocial}}</p>
+          <p>Direccion: {{domicilio}}</p>
+          <p>C.U.I.T.: {{cuit}}</p>
+          <p>{{regimenTributario}}</p>
+          <p>IIBB: {{iibb}}</p>
+          <p>Inicio de actividad: {{inicioActividad}}</p>
+        </td>
+      </tr>
+      <tr>
+        <td class="border-top padding-t-3 padding-b-3">
+          <p class="text-center text-lg">FACTURA A</p>
+          <!-- 
+          <p class="text-center">Codigo {{cbteTipo}}</p> -->
+          <p>P.V: {{ptoVta}}</p>
+          <p>Nro: {{numeroFactura}}</p>
+          <p>Fecha: {{fecha}}</p>
+          <p>Concepto: {{concepto}}</p>
+        </td>
+      </tr>
+      <tr>
+        <td class="border-top padding-t-3 padding-b-3">
+          <p>{{condicionIVA}}</p>
+        </td>
+      </tr>
+      <tr>
+        <td class="border-top padding-t-3 padding-b-3">
+          <div>
+            <table>
+              {{productRows}}
+            </table>
+          </div>
+        </td>
+      </tr>
+      <tr>
+        <td class="border-top padding-t-3 padding-b-3">
+          <div>
+            <table>
+              <tr>
+                <td>Subtotal</td>
+                <td>{{ImpNeto}}</td>
+              </tr>
+              <tr>
+                <td>IVA 21 %</td>
+                <td>{{ImpIVA}}</td>
+              </tr>
+              <tr>
+                <td>Descuento %</td>
+                <td>{{discount}}</td>
+              </tr>
+            </table>
+          </div>
+        </td>
+      </tr>
+      <tr>
+        <td class="border-top padding-t-3 padding-b-3">
+          <div>
+            <table>
+              <tr>
+                <td>TOTAL</td>
+                <td>{{ImpTotal}}</td>
+              </tr>
+            </table>
+          </div>
+        </td>
+      </tr>
+      <tr>
+        <td class="border-top padding-t-3">
+          <p>CAE: {{CAE}}</p>
+          <p>Vto: {{Vencimiento}}</p>
+        </td>
+      </tr>
+      <tr class="text-center">
+        <td>
+          <img src="{{urlQr}}" />
+        </td>
+      </tr>
+    </table>
+  </body>
+</html>
